Split AST analyzer Go helpers
